+++ bruges/reflection/reflection.py
@@ -9,34 +9,66 @@ from functools import wraps
 from collections import namedtuple
 
 import numpy as np
-from numpy import tan, sin, cos
 
 from bruges.rockphysics import moduli
 from bruges.rockphysics import anisotropy
 from bruges.util import deprecated
 
 
-def reflectivity(vp, vs, rho, theta, method='zoeppritz_rpp'):
+def acoustic_reflectivity(vp, rho):
     """
-    Compute 'upper' and 'lower' intervals from the three provided arrays,
-    then pass the result to the specified method to compute reflection
+    The acoustic reflectivity, given Vp and RHOB logs.
+
+    Args:
+        vp (ndarray): The P-wave velocity.
+        rho (ndarray): The bulk density.
+
+    Returns:
+        ndarray: The reflectivity coefficient series.
+    """
+    upper = vp[:-1] * rho[:-1]
+    lower = vp[1:] * rho[1:]
+    return (lower - upper) / (lower + upper)
+
+
+def reflectivity(vp, vs, rho, theta=0, method='zoeppritz_rpp'):
+    """
+    Offset reflectivity, given Vp, Vs, rho, and offset.
+
+    Computes 'upper' and 'lower' intervals from the three provided arrays,
+    then passes the result to the specified method to compute reflection
     coefficients.
 
-    :param vp: The P-wave velocity 1D array.
-    :param vs: The S-wave velocity 1D array.
-    :param rho: The density 1D array.
-    :param theta: Incidence angle, a scalar or 1D array [degrees].
-    :param method: Which method to use:
+    For acoustic reflectivity, either use the `acoustic_reflectivity()`
+    function, or call `reflectivity()` passing any log as Vs, e.g. just give
+    the Vp log twice (it won't be used anyway):
 
-        - 'zoeppritz_rpp': zoeppritz_rpp,
-        - 'akirichards': akirichards,
-        - 'akirichards_alt': akirichards_alt,
-        - 'fatti': fatti,
-        - 'shuey': shuey,
-        - 'bortfeld': bortfeld,
-        - 'hilterman': hilterman,
+        reflectivity(vp, vp, rho)
 
-    :returns: the result of running the specified method on the data.
+    For anisotropic reflectivity, use either `anisotropy.blangy()` or
+    `anisotropy.ruger()`.
+
+    Args:
+        vp (ndarray): The P-wave velocity; float or 1D array length m.
+        vs (ndarray): The S-wave velocity; float or 1D array length m.
+        rho (ndarray): The density; float or 1D array length m.
+        theta (ndarray): The incidence angle; float or 1D array length n.
+        method (str): The reflectivity equation to use; one of:
+
+                - 'zoeppritz_rpp': zoeppritz_rpp,
+                - 'akirichards': akirichards,
+                - 'akirichards_alt': akirichards_alt,
+                - 'fatti': fatti,
+                - 'shuey': shuey,
+                - 'bortfeld': bortfeld,
+                - 'hilterman': hilterman,
+
+    Returns:
+        ndarray. The result of running the specified method on the inputs.
+            Will be a float (for float inputs and one angle), a 1 x n array
+            (for float inputs and an array of angles), a 1 x m-1 array (for
+            float inputs and one angle), or an m-1 x n array (for array inputs
+            and an array of angles).
     """
     methods = {
         'zoeppritz_rpp': zoeppritz_rpp,
@@ -51,9 +83,11 @@ def reflectivity(vp, vs, rho, theta, method='zoeppritz_rpp'):
     vp = np.array(vp).astype(float)
     vs = np.array(vs).astype(float)
     rho = np.array(rho).astype(float)
+
     vp1, vp2 = vp[:-1], vp[1:]
     vs1, vs2 = vs[:-1], vs[1:]
     rho1, rho2 = rhob[:-1], rhob[1:]
+
     return func(vp1, vs1, rho1, vp2, vs2, rho2, theta)
 
 
@@ -61,130 +95,142 @@ def vectorize(func):
     """
     Decorator to make sure the inputs are arrays. We also add a dimension
     to theta to make the functions work in an 'outer product' way.
+
+    Takes a reflectivity function requiring Vp, Vs, and RHOB for 2 rocks
+    (upper and lower), plus incidence angle theta, plus kwargs. Returns
+    that function with the arguments transformed to ndarrays.
     """
     @wraps(func)
     def wrapper(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, **kwargs):
         vp1 = np.array(vp1).astype(float)
-        vp2 = np.array(vp2).astype(float)
         vs1 = np.array(vs1).astype(float)
-        vs2 = np.array(vs2).astype(float)
         rho1 = np.array(rho1).astype(float)
+        vp2 = np.array(vp2).astype(float)
+        vs2 = np.array(vs2).astype(float)
         rho2 = np.array(rho2).astype(float)
         theta1 = np.array(theta1).reshape((-1, 1))
         return func(vp1, vs1, rho1, vp2, vs2, rho2, theta1, **kwargs)
     return wrapper
 
 
-def scattering_matrix(vp1, vs1, rho1, vp0, vs0, rho0, theta1=0):
-    '''
+def scattering_matrix(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0):
+    """
     Full Zoeppritz solution, considered the definitive solution.
     Calculates the angle dependent p-wave reflectivity of an interface
     between two mediums.
 
     Originally written by: Wes Hamlyn, vectorized by Agile.
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
+    Args:
+        vp1 (float): The upper P-wave velocity.
+        vs1 (float): The upper S-wave velocity.
+        rho1 (float): The upper layer's density.
+        vp2 (float): The lower P-wave velocity.
+        vs2 (float): The lower S-wave velocity.
+        rho2 (float): The lower layer's density.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
 
-    :param vp0: The p-wave velocity of the lower medium.
-    :param vs0: The s-wave velocity of the lower medium.
-    :param rho0: The density of the lower medium.
-
-    :param theta1: A scalar  [degrees].
-
-    :returns: a 4x4 array representing the scattering matrix
-                at the incident angle theta1.
-    '''
+    Returns:
+        ndarray. The exact Zoeppritz solution for all modes at the interface.
+            A 4x4 array representing the scattering matrix at the incident
+            angle theta1.
+    """
     theta1 = np.radians(np.array(theta1))
     if theta1.size == 1:
         theta1 = np.expand_dims(theta1, axis=1)
 
-    p = sin(theta1) / vp1  # Ray parameter.
+    p = np.sin(theta1) / vp1  # Ray parameter.
 
     # Calculate reflection & transmission angles for Zoeppritz.
-    theta2 = np.arcsin(p * vp0)  # Trans. angle of P-wave.
+    theta2 = np.arcsin(p * vp2)  # Trans. angle of P-wave.
     phi1 = np.arcsin(p * vs1)    # Refl. angle of converted S-wave.
-    phi2 = np.arcsin(p * vs0)    # Trans. angle of converted S-wave.
+    phi2 = np.arcsin(p * vs2)    # Trans. angle of converted S-wave.
 
     # Matrix form of Zoeppritz Equations... M & N are matrices.
-    M = np.array([[-sin(theta1), -cos(phi1), sin(theta2), cos(phi2)],
-                  [cos(theta1), -sin(phi1), cos(theta2), -sin(phi2)],
-                  [2 * rho1 * vs1 * sin(phi1) * cos(theta1),
-                   rho1 * vs1 * (1 - 2 * sin(phi1) ** 2),
-                   2 * rho0 * vs0 * sin(phi2) * cos(theta2),
-                   rho0 * vs0 * (1 - 2 * sin(phi2) ** 2)],
-                  [-rho1 * vp1 * (1 - 2 * sin(phi1) ** 2),
-                   rho1 * vs1 * sin(2 * phi1),
-                   rho0 * vp0 * (1 - 2 * sin(phi2) ** 2),
-                   -rho0 * vs0 * sin(2 * phi2)]], dtype='float')
+    M = np.array([[-np.sin(theta1),-np.cos(phi1),np.sin(theta2), np.cos(phi2)],
+                  [np.cos(theta1),-np.sin(phi1),np.cos(theta2), -np.sin(phi2)],
+                  [2 * rho1 * vs1 * np.sin(phi1) * np.cos(theta1),
+                   rho1 * vs1 * (1 - 2 * np.sin(phi1) ** 2),
+                   2 * rho2 * vs2 * np.sin(phi2) * np.cos(theta2),
+                   rho2 * vs2 * (1 - 2 * np.sin(phi2) ** 2)],
+                  [-rho1 * vp1 * (1 - 2 * np.sin(phi1) ** 2),
+                   rho1 * vs1 * np.sin(2 * phi1),
+                   rho2 * vp2 * (1 - 2 * np.sin(phi2) ** 2),
+                   -rho2 * vs2 * np.sin(2 * phi2)]], dtype='float')
 
-    N = np.array([[sin(theta1), cos(phi1), -sin(theta2), -cos(phi2)],
-                  [cos(theta1), -sin(phi1), cos(theta2), -sin(phi2)],
-                  [2 * rho1 * vs1 * sin(phi1) * cos(theta1),
-                   rho1 * vs1 * (1 - 2 * sin(phi1) ** 2),
-                   2 * rho0 * vs0 * sin(phi2) * cos(theta2),
-                   rho0 * vs0 * (1 - 2 * sin(phi2) ** 2)],
-                  [rho1 * vp1 * (1 - 2 * sin(phi1) ** 2),
-                   -rho1 * vs1 * sin(2 * phi1),
-                   - rho0 * vp0 * (1 - 2 * sin(phi2) ** 2),
-                   rho0 * vs0 * sin(2 * phi2)]], dtype='float')
+    N = np.array([[np.sin(theta1),np.cos(phi1),-np.sin(theta2), -np.cos(phi2)],
+                  [np.cos(theta1),-np.sin(phi1),np.cos(theta2), -np.sin(phi2)],
+                  [2 * rho1 * vs1 * np.sin(phi1) * np.cos(theta1),
+                   rho1 * vs1 * (1 - 2 * np.sin(phi1) ** 2),
+                   2 * rho2 * vs2 * np.sin(phi2) * np.cos(theta2),
+                   rho2 * vs2 * (1 - 2 * np.sin(phi2) ** 2)],
+                  [rho1 * vp1 * (1 - 2 * np.sin(phi1) ** 2),
+                   -rho1 * vs1 * np.sin(2 * phi1),
+                   - rho2 * vp2 * (1 - 2 * np.sin(phi2) ** 2),
+                   rho2 * vs2 * np.sin(2 * phi2)]], dtype='float')
 
     A = np.linalg.inv(np.rollaxis(M, 2))
     Z = np.matmul(A, np.rollaxis(N, -1))
 
-    return np.rollaxis(Z, 0, 3)
+    return np.rollaxis(Z, 0, 3).T
 
 
-def zoeppritz_element(vp1, vs1, rho1, vp0, vs0, rho0, theta1=0, element='PdPu'):
+def zoeppritz_element(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, element='PdPu'):
     """
     Returns any mode reflection coefficients from the Zoeppritz
     scattering matrix. Pass in the mode as element, e.g. 'PdSu' for PS.
 
     Wraps scattering_matrix().
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
+    Args:
+        vp1 (float): The upper P-wave velocity.
+        vs1 (float): The upper S-wave velocity.
+        rho1 (float): The upper layer's density.
+        vp2 (float): The lower P-wave velocity.
+        vs2 (float): The lower S-wave velocity.
+        rho2 (float): The lower layer's density.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
+        element (str): The name of the element to return, must be one of:
+            'PdPu', 'SdPu', 'PuPu', 'SuPu', 'PdSu', 'SdSu', 'PuSu', 'SuSu',
+            'PdPd', 'SdPd', 'PuPd', 'SuPd', 'PdSd', 'SdSd', 'PuSd', 'SuSd'.
 
-    :param vp0: The p-wave velocity of the lower medium.
-    :param vs0: The s-wave velocity of the lower medium.
-    :param rho0: The density of the lower medium.
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
+    Returns:
+        ndarray. Array length n of the exact Zoeppritz solution for the
+            specified modes at the interface, at the incident angle theta1.
     """
     elements = np.array([['PdPu', 'SdPu', 'PuPu', 'SuPu'],
                          ['PdSu', 'SdSu', 'PuSu', 'SuSu'],
                          ['PdPd', 'SdPd', 'PuPd', 'SuPd'],
                          ['PdSd', 'SdSd', 'PuSd', 'SuSd']])
 
-    Z = scattering_matrix(vp1, vs1, rho1, vp0, vs0, rho0, theta1)
+    Z = scattering_matrix(vp1, vs1, rho1, vp2, vs2, rho2, theta1).T
 
     return np.squeeze(Z[np.where(elements == element)])
 
 
-def zoeppritz(vp1, vs1, rho1, vp0, vs0, rho0, theta1=0):
-    '''
+def zoeppritz(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0):
+    """
     Returns the PP reflection coefficients from the Zoeppritz
     scattering matrix. Wraps zoeppritz_element().
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
+    Args:
+        vp1 (float): The upper P-wave velocity.
+        vs1 (float): The upper S-wave velocity.
+        rho1 (float): The upper layer's density.
+        vp2 (float): The lower P-wave velocity.
+        vs2 (float): The lower S-wave velocity.
+        rho2 (float): The lower layer's density.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
 
-    :param vp0: The p-wave velocity of the lower medium.
-    :param vs0: The s-wave velocity of the lower medium.
-    :param rho0: The density of the lower medium.
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
-    '''
-    return zoeppritz_element(vp1, vs1, rho1, vp0, vs0, rho0, theta1, 'PdPu')
+    Returns:
+        ndarray. Array length n of the exact Zoeppritz solution for the
+            specified modes at the interface, at the incident angle theta1.
+    """
+    return zoeppritz_element(vp1, vs1, rho1, vp2, vs2, rho2, theta1, 'PdPu')
 
 
 @vectorize
-def zoeppritz_rpp(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
+def zoeppritz_rpp(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0):
     """
     Exact Zoeppritz from expression.
 
@@ -193,15 +239,21 @@ def zoeppritz_rpp(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
 
     Dvorkin et al. (2014). Seismic Reflections of Rock Properties. Cambridge.
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
+    Args:
+        vp1 (ndarray): The upper P-wave velocity; float or 1D array length m.
+        vs1 (ndarray): The upper S-wave velocity; float or 1D array length m.
+        rho1 (ndarray): The upper layer's density; float or 1D array length m.
+        vp2 (ndarray): The lower P-wave velocity; float or 1D array length m.
+        vs2 (ndarray): The lower S-wave velocity; float or 1D array length m.
+        rho2 (ndarray): The lower layer's density; float or 1D array length m.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
 
-    :param vp0: The p-wave velocity of the lower medium.
-    :param vs0: The s-wave velocity of the lower medium.
-    :param rho0: The density of the lower medium.
-
-    :returns: a matrix of length `len(vp1)` whose rows are of length `len(vp[RHOB])`.
+    Returns:
+        ndarray. The exact Zoeppritz solution for P-P reflectivity at the
+            interface. Will be a float (for float inputs and one angle), a
+            1 x n array (for float inputs and an array of angles), a 1 x m
+            array (for float inputs and one angle), or an m x n array (for
+            array inputs and an array of angles).
     """
     theta1 = np.radians(theta1)
 
@@ -231,31 +283,36 @@ def zoeppritz_rpp(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
 @vectorize
 def akirichards(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     """
-    This is the formulation from Avseth et al.,
-    Quantitative seismic interpretation,
-    Cambridge University Press, 2006. Adapted for a 4-term formula.
-    See http://subsurfwiki.org/wiki/Aki-Richards_equation
+    The Aki-Richards approximation to the reflectivity.
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
+    This is the formulation from Avseth et al., _Quantitative seismic
+    interpretation_, Cambridge University Press, 2006. Adapted for a 4-term
+    formula. See http://subsurfwiki.org/wiki/Aki-Richards_equation.
 
-    :param vp2: The p-wave velocity of the lower medium.
-    :param vs2: The s-wave velocity of the lower medium.
-    :param rho2: The density of the lower medium.
+    Args:
+        vp1 (ndarray): The upper P-wave velocity; float or 1D array length m.
+        vs1 (ndarray): The upper S-wave velocity; float or 1D array length m.
+        rho1 (ndarray): The upper layer's density; float or 1D array length m.
+        vp2 (ndarray): The lower P-wave velocity; float or 1D array length m.
+        vs2 (ndarray): The lower S-wave velocity; float or 1D array length m.
+        rho2 (ndarray): The lower layer's density; float or 1D array length m.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
+        terms (bool): Whether or not to return a tuple of the terms of the
+            equation. The first term is the acoustic impedance.
 
-    :param theta1: An array of incident angles to use for reflectivity
-                   calculation [degrees].
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle
+    Returns:
+        ndarray. The Aki-Richards approximation for P-P reflectivity at the
+            interface. Will be a float (for float inputs and one angle), a
+            1 x n array (for float inputs and an array of angles), a 1 x m
+            array (for float inputs and one angle), or an m x n array (for
+            array inputs and an array of angles).
     """
     theta1 = np.radians(theta1)
     if (np.nan_to_num(theta1) > np.pi/2.).any():
         raise ValueError("Incidence angle theta1 must be less than 90 deg.")
 
     # critical_angle = arcsin(vp1/vp2)
-    theta2 = np.arcsin(vp2/vp1*sin(theta1))
+    theta2 = np.arcsin(vp2/vp1*np.sin(theta1))
     drho = rho2-rho1
     dvp = vp2-vp1
     dvs = vs2-vs1
@@ -294,24 +351,28 @@ def akirichards_alt(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     This is another formulation of the Aki-Richards solution.
     See http://subsurfwiki.org/wiki/Aki-Richards_equation
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
+    Args:
+        vp1 (ndarray): The upper P-wave velocity; float or 1D array length m.
+        vs1 (ndarray): The upper S-wave velocity; float or 1D array length m.
+        rho1 (ndarray): The upper layer's density; float or 1D array length m.
+        vp2 (ndarray): The lower P-wave velocity; float or 1D array length m.
+        vs2 (ndarray): The lower S-wave velocity; float or 1D array length m.
+        rho2 (ndarray): The lower layer's density; float or 1D array length m.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
+        terms (bool): Whether or not to return a tuple of the terms of the
+            equation. The first term is the acoustic impedance.
 
-    :param vp2: The p-wave velocity of the lower medium.
-    :param vs2: The s-wave velocity of the lower medium.
-    :param rho2: The density of the lower medium.
-
-    :param theta1: An array of incident angles to use for reflectivity
-                   calculation [degrees].
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
+    Returns:
+        ndarray. The Aki-Richards approximation for P-P reflectivity at the
+            interface. Will be a float (for float inputs and one angle), a
+            1 x n array (for float inputs and an array of angles), a 1 x m
+            array (for float inputs and one angle), or an m x n array (for
+            array inputs and an array of angles).
     """
     theta1 = np.radians(theta1)
 
     # critical_angle = arcsin(vp1/vp2)
-    theta2 = np.arcsin(vp2/vp1*sin(theta1))
+    theta2 = np.arcsin(vp2/vp1*np.sin(theta1))
     drho = rho2-rho1
     dvp = vp2-vp1
     dvs = vs2-vs1
@@ -322,8 +383,8 @@ def akirichards_alt(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
 
     # Compute the three terms
     term1 = 0.5 * (dvp/vp + drho/rho)
-    term2 = (0.5*dvp/vp-2*(vs/vp)**2*(drho/rho+2*dvs/vs)) * sin(theta)**2
-    term3 = 0.5 * dvp/vp * (tan(theta)**2 - sin(theta)**2)
+    term2 = (0.5*dvp/vp-2*(vs/vp)**2*(drho/rho+2*dvs/vs)) * np.sin(theta)**2
+    term3 = 0.5 * dvp/vp * (np.tan(theta)**2 - np.sin(theta)**2)
 
     if terms:
         fields = ['term1', 'term2', 'term3']
@@ -339,23 +400,27 @@ def akirichards_alt(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
 @vectorize
 def fatti(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     """
-    Compute reflectivities with Fatti's formulation of the
-    Aki-Richards equation, which does not account for the
-    critical angle. Fatti et al (1994), Geophysics 59 (9).
+    Compute reflectivities with Fatti's formulation of the Aki-Richards
+    equation, which does not account for the critical angle. See Fatti et al.
+    (1994), Geophysics 59 (9).
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
+    Args:
+        vp1 (ndarray): The upper P-wave velocity; float or 1D array length m.
+        vs1 (ndarray): The upper S-wave velocity; float or 1D array length m.
+        rho1 (ndarray): The upper layer's density; float or 1D array length m.
+        vp2 (ndarray): The lower P-wave velocity; float or 1D array length m.
+        vs2 (ndarray): The lower S-wave velocity; float or 1D array length m.
+        rho2 (ndarray): The lower layer's density; float or 1D array length m.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
+        terms (bool): Whether or not to return a tuple of the terms of the
+            equation. The first term is the acoustic impedance.
 
-    :param vp2: The p-wave velocity of the lower medium.
-    :param vs2: The s-wave velocity of the lower medium.
-    :param rho2: The density of the lower medium.
-
-    :param theta1: An array of incident angles to use for reflectivity
-                   calculation [degrees].
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
+    Returns:
+        ndarray. The Fatti approximation for P-P reflectivity at the
+            interface. Will be a float (for float inputs and one angle), a
+            1 x n array (for float inputs and an array of angles), a 1 x m
+            array (for float inputs and one angle), or an m x n array (for
+            array inputs and an array of angles).
     """
     theta1 = np.radians(theta1)
 
@@ -368,9 +433,9 @@ def fatti(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     d = drho/rho
 
     # Compute the three terms
-    term1 = (1 + tan(theta1)**2) * dip
-    term2 = -8 * (vs/vp)**2 * dis * sin(theta1)**2
-    term3 = -1 * (0.5 * tan(theta1)**2 - 2 * (vs/vp)**2 * sin(theta1)**2) * d
+    term1 = (1 + np.tan(theta1)**2) * dip
+    term2 = -8 * (vs/vp)**2 * dis * np.sin(theta1)**2
+    term3 = -1 * (0.5 * np.tan(theta1)**2 - 2 * (vs/vp)**2 * np.sin(theta1)**2) * d
 
     if terms:
         fields = ['term1', 'term2', 'term3']
@@ -391,21 +456,25 @@ def shuey(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0,
     Compute Shuey approximation with 3 terms.
     http://subsurfwiki.org/wiki/Shuey_equation
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
-    :param vp2: The p-wave velocity of the lower medium.
-    :param vs2: The s-wave velocity of the lower medium.
-    :param rho2: The density of the lower medium.
-    :param theta1: An array of incident angles to use for reflectivity
-                   calculation [degrees].
-    :param terms: bool. Whether to return a tuple of the 3 individual terms.
-    :param return_gradient: bool. Whether to return a tuple of the intercept
-                            and gradient (i.e. the second term divided by
-                            sin^2(theta).
+    Args:
+        vp1 (ndarray): The upper P-wave velocity; float or 1D array length m.
+        vs1 (ndarray): The upper S-wave velocity; float or 1D array length m.
+        rho1 (ndarray): The upper layer's density; float or 1D array length m.
+        vp2 (ndarray): The lower P-wave velocity; float or 1D array length m.
+        vs2 (ndarray): The lower S-wave velocity; float or 1D array length m.
+        rho2 (ndarray): The lower layer's density; float or 1D array length m.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
+        terms (bool): Whether or not to return a tuple of the terms of the
+            equation. The first term is the acoustic impedance.
+        return_gradient (bool): Whether to return a tuple of the intercept
+            and gradient (i.e. the second term divided by sin^2(theta)).
 
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
+    Returns:
+        ndarray. The Aki-Richards approximation for P-P reflectivity at the
+            interface. Will be a float (for float inputs and one angle), a
+            1 x n array (for float inputs and an array of angles), a 1 x m
+            array (for float inputs and one angle), or an m x n array (for
+            array inputs and an array of angles).
     """
     theta1 = np.radians(theta1)
 
@@ -443,7 +512,8 @@ def shuey(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0,
 @deprecated('Please use shuey() instead.')
 def shuey2(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0):
     """
-    Compute Shuey approximation with 2 terms.
+    Compute Shuey approximation with 2 terms. Wraps `shuey()`. Deprecated,
+    use `shuey()` instead.
     """
     r, g, _ = shuey(vp1, vs1, rho1, vp2, vs2, rho2, theta1=theta1, terms=True)
     return r + g
@@ -452,50 +522,10 @@ def shuey2(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0):
 @deprecated('Please use shuey() instead.')
 def shuey3(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     """
-    Compute Shuey approximation with 3 terms.
+    Compute Shuey approximation with 3 terms. Wraps `shuey()`. Deprecated,
+    use `shuey()` instead.
     """
     return shuey(vp1, vs1, rho1, vp2, vs2, rho2, theta1=theta1)
-
-
-@deprecated('Please use bortfeld() instead.')
-def bortfeld2(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
-    """
-    The 2-term Bortfeld approximation for ava analysis.
-
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
-
-    :param vp2: The p-wave velocity of the lower medium.
-    :param vs2: The s-wave velocity of the lower medium.
-    :param rho2: The density of the lower medium.
-
-    :param theta1: An array of incident angles to use for reflectivity
-                   calculation [degrees].
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
-    """
-    theta1 = np.radians(theta1)
-    # Bortfeld only needs one extra parameter
-    theta2 = np.arcsin(vp2/vp1*np.sin(theta1))  # radians
-
-    # This breaks if theta = 90 deg
-    term1 = 0.5 * np.log((vp2*rho2*cos(theta1)) / (vp1*rho1*cos(theta2)))
-
-    svp2 = (np.sin(theta1)/vp1)**2
-    dvs2 = (vs1**2-vs2**2)
-    term2 = svp2 * dvs2 * (2+np.log(rho2/rho1)/np.log(vs2/vs1))
-
-    if terms:
-        return term1, term2
-    else:
-        return (term1 + term2)
-
-
-@deprecated('Please use bortfeld() instead.')
-def bortfeld3(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
-    return bortfeld(vp1, vs1, rho1, vp2, vs2, rho2, theta1=theta1)
 
 
 @vectorize
@@ -504,19 +534,23 @@ def bortfeld(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     Compute Bortfeld approximation with three terms.
     http://sepwww.stanford.edu/public/docs/sep111/marie2/paper_html/node2.html
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
+    Args:
+        vp1 (ndarray): The upper P-wave velocity; float or 1D array length m.
+        vs1 (ndarray): The upper S-wave velocity; float or 1D array length m.
+        rho1 (ndarray): The upper layer's density; float or 1D array length m.
+        vp2 (ndarray): The lower P-wave velocity; float or 1D array length m.
+        vs2 (ndarray): The lower S-wave velocity; float or 1D array length m.
+        rho2 (ndarray): The lower layer's density; float or 1D array length m.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
+        terms (bool): Whether or not to return a tuple of the terms of the
+            equation. The first term is the acoustic impedance.
 
-    :param vp2: The p-wave velocity of the lower medium.
-    :param vs2: The s-wave velocity of the lower medium.
-    :param rho2: The density of the lower medium.
-
-    :param theta1: An array of incident angles to use for reflectivity
-                   calculation [degrees].
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
+    Returns:
+        ndarray. The 3-term Bortfeld approximation for P-P reflectivity at the
+            interface. Will be a float (for float inputs and one angle), a
+            1 x n array (for float inputs and an array of angles), a 1 x m
+            array (for float inputs and one angle), or an m x n array (for
+            array inputs and an array of angles).
     """
     theta1 = np.radians(theta1)
 
@@ -545,6 +579,48 @@ def bortfeld(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
         return np.squeeze(term1 + term2 + term3).T
 
 
+@deprecated('Please use bortfeld() instead.')
+def bortfeld2(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
+    """
+    The 2-term Bortfeld approximation for ava analysis. Wraps `shuey()`.
+    Deprecated, use `bortfeld()` instead.
+
+    Args:
+        vp1 (ndarray): The upper P-wave velocity; float or 1D array length m.
+        vs1 (ndarray): The upper S-wave velocity; float or 1D array length m.
+        rho1 (ndarray): The upper layer's density; float or 1D array length m.
+        vp2 (ndarray): The lower P-wave velocity; float or 1D array length m.
+        vs2 (ndarray): The lower S-wave velocity; float or 1D array length m.
+        rho2 (ndarray): The lower layer's density; float or 1D array length m.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
+        terms (bool): Whether or not to return a tuple of the terms of the
+            equation. The first term is the acoustic impedance.
+
+    Returns:
+        ndarray. The 2-term Bortfeld approximation for P-P reflectivity at the
+            interface. Will be a float (for float inputs and one angle), a
+            1 x n array (for float inputs and an array of angles), a 1 x m
+            array (for float inputs and one angle), or an m x n array (for
+            array inputs and an array of angles).
+    """
+    theta1 = np.radians(theta1)
+    theta2 = np.arcsin(vp2/vp1*np.sin(theta1))
+    term1 = 0.5 * np.log((vp2*rho2*np.cos(theta1)) / (vp1*rho1*np.cos(theta2)))
+    svp2 = (np.sin(theta1)/vp1)**2
+    dvs2 = (vs1**2-vs2**2)
+    term2 = svp2 * dvs2 * (2+np.log(rho2/rho1)/np.log(vs2/vs1))
+
+    if terms:
+        return term1, term2
+    else:
+        return (term1 + term2)
+
+
+@deprecated('Please use bortfeld() instead.')
+def bortfeld3(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
+    return bortfeld(vp1, vs1, rho1, vp2, vs2, rho2, theta1=theta1)
+
+
 @vectorize
 def hilterman(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     """
@@ -554,19 +630,23 @@ def hilterman(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
     According to Dvorkin: "arguably the simplest and a very convenient
     [approximation]." At least for small angles and small contrasts.
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
+    Args:
+        vp1 (ndarray): The upper P-wave velocity; float or 1D array length m.
+        vs1 (ndarray): The upper S-wave velocity; float or 1D array length m.
+        rho1 (ndarray): The upper layer's density; float or 1D array length m.
+        vp2 (ndarray): The lower P-wave velocity; float or 1D array length m.
+        vs2 (ndarray): The lower S-wave velocity; float or 1D array length m.
+        rho2 (ndarray): The lower layer's density; float or 1D array length m.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
+        terms (bool): Whether or not to return a tuple of the terms of the
+            equation. The first term is the acoustic impedance.
 
-    :param vp2: The p-wave velocity of the lower medium.
-    :param vs2: The s-wave velocity of the lower medium.
-    :param rho2: The density of the lower medium.
-
-    :param theta1: An array of incident angles to use for reflectivity
-                   calculation [degrees].
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
+    Returns:
+        ndarray. The Hilterman approximation for P-P reflectivity at the
+            interface. Will be a float (for float inputs and one angle), a
+            1 x n array (for float inputs and an array of angles), a 1 x m
+            array (for float inputs and one angle), or an m x n array (for
+            array inputs and an array of angles).
     """
     theta1 = np.radians(theta1)
 
@@ -591,33 +671,27 @@ def hilterman(vp1, vs1, rho1, vp2, vs2, rho2, theta1=0, terms=False):
 
 def blangy(vp1, vs1, rho1, vp2, vs2, rho2,
            d1=0, e1=0, d2=0, e2=0,
-           theta1=0, terms=False):
+           theta1=0):
     """Implements the Blangy equation with the same interface as the other
-    reflectivity equations. Wraps bruges.anisotropy.blangy().
+    reflectivity equations. Wraps bruges.anisotropy.blangy(), which you may
+    prefer to use directly.
 
-    Note that the anisotropic parameters come after the other rock properties,
-    and all default to zero.
+    Args:
+        vp1 (ndarray): The upper P-wave velocity; float or 1D array length m.
+        vs1 (ndarray): The upper S-wave velocity; float or 1D array length m.
+        rho1 (ndarray): The upper layer's density; float or 1D array length m.
+        vp2 (ndarray): The lower P-wave velocity; float or 1D array length m.
+        vs2 (ndarray): The lower S-wave velocity; float or 1D array length m.
+        rho2 (ndarray): The lower layer's density; float or 1D array length m.
+        d1 (ndarray): The upper delta; float or 1D array length m.
+        e1 (ndarray): The upper epsilon; float or 1D array length m.
+        d2 (ndarray): The lower delta; float or 1D array length m.
+        e2 (ndarray): The lower epsilon; float or 1D array length m.
+        theta1 (ndarray): The incidence angle; float or 1D array length n.
 
-    :param vp1: The p-wave velocity of the upper medium.
-    :param vs1: The s-wave velocity of the upper medium.
-    :param rho1: The density of the upper medium.
-    :param vp2: The p-wave velocity of the lower medium.
-    :param vs2: The s-wave velocity of the lower medium.
-    :param rho2: The density of the lower medium.
-    :param d1: Thomsen's delta for the upper medium.
-    :param e1: Thomsen's epsilon for the upper medium.
-    :param d2: Thomsen's delta for the upper medium.
-    :param e2: Thomsen's epsilon for the upper medium.
-    :param theta1: An array of incident angles to use for reflectivity
-                   calculation [degrees].
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
-    :param theta1: An array of incident angles to use for reflectivity
-                   calculation [degrees].
-
-    :returns: a vector of len(theta1) containing the reflectivity
-             value corresponding to each angle.
+    Returns:
+        ndarray. The Blangy approximation for P-P reflectivity at the
+            interface. Wraps `anisotropy.blangy()`.
     """
     _, anisotropic = anisotropy.blangy(vp1, vs1, rho1, d1, e1,  # UPPER
                                        vp2, vs2, rho2, d2, e2,  # LOWER